Make it one fragment to holds all BrandsFragment

+++ app/src/main/java/com/example/laza/data/Product.kt
@@ -7,7 +7,7 @@ data class Product(
     val price: Float,
     val offerPercentage: Float? = null,
     val description: String? = null,
-    val colors: List<Int>? = null,
+    val colors: List<Double>? = null,
     val sizes: List<String>? = null,
     val images: List<String>
 ){
